requirements.txt
# Core libraries
RPi.GPIO
speechrecognition
pyaudio
transformers
torch
accelerate
numpy
pillow

# Networking
socket

# System utilities
espeak

# For Pi camera (optional)
opencv-python

# Optional (if Pi uses Python 3.9+, accelerate will handle backend)
